% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ud.functions.R
\name{extract.unit.parts}
\alias{extract.unit.parts}
\title{Extract the text and value part from the unit text}
\usage{
extract.unit.parts(unitdef, assumed.value = 1)
}
\arguments{
\item{unitdef}{A unit definition in the form of "optional blank, 
optional floating point nonnegative number, optional blank, unit 
text, optional blank".  Blanks (e.g. spaces, tabs, etc.) are 
ignored.}

\item{assumed.value}{The value to assume if the optional floating 
point nonnegative number is not given in the \code{unitdef}.}
}
\value{
A list with two elements:  \code{value} and \code{text}.
}
\description{
Extract the text and value part from the unit text
}
\note{
This function is called from ud.set.conversion
}
\author{
William Denney \email{wdenney@humanpredictions.com}
}

