% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ud.functions.R
\name{ud.add.system}
\alias{ud.add.system}
\title{Add a unit system}
\usage{
ud.add.system(file.name, system.name)
}
\arguments{
\item{file.name}{The XML file to read for unit information.  If 
\code{file.name} is blank, the system default xml is used.}

\item{system.name}{The unit system name to use.  If system.name is
null the the name of the system is set to "default".}
}
\description{
Add a unit system
}
\author{
Swechhya Bista \email{Swechhya.Bista@cytel.com}
}
\seealso{
\code{\link{ud.list.systems}}, \code{\link{ud.free.system}}
}

