% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ud.functions.R
\name{ud.free.system}
\alias{ud.free.system}
\title{Delete the system given by the user}
\usage{
ud.free.system(system.name)
}
\arguments{
\item{system.name}{The unit system name to use.}
}
\value{
\code{TRUE} if system.name is successfully deleted, otherwise
  returns an error.
}
\description{
Delete the system given by the user
}
\author{
Swechhya Bista \email{Swechhya.Bista@cytel.com}
}

