% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ud.functions.R
\name{ud.have.unit.system}
\alias{ud.have.unit.system}
\title{Check if a system is loaded.}
\usage{
ud.have.unit.system(system.name)
}
\arguments{
\item{system.name}{The unit system name to check.}
}
\value{
\code{TRUE} if the system specified by system.name is
  present, otherwise \code{FALSE}.
}
\description{
Check if a system is loaded.
}
\examples{
ud.have.unit.system("default") # TRUE
ud.have.unit.system("notloaded") # FALSE
}
\author{
Swechhya Bista \email{Swechhya.Bista@cytel.com}
}
\seealso{
\code{\link{ud.add.system}}
}

