% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ud.functions.R
\name{ud.list.systems}
\alias{ud.list.systems}
\title{List the loaded systems}
\usage{
ud.list.systems()
}
\value{
A character vector containing the system names.  If no 
  systems have been loaded it returns \code{NULL}.
}
\description{
List the loaded systems
}
\author{
Swechhya Bista \email{Swechhya.Bista@cytel.com}
}
\seealso{
\code{\link{ud.add.system}}, \code{\link{ud.free.system}}
}

