% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ud.functions.R
\name{ud.remove.unit}
\alias{ud.remove.unit}
\title{Remove a unit from a loaded system}
\usage{
ud.remove.unit(unit.name, system.name = "default")
}
\arguments{
\item{unit.name}{The unit name to remove from a system}

\item{system.name}{The unit system name to use.}
}
\value{
\code{TRUE} if the unit is successfully removed, otherwise
  returns an error.
}
\description{
Remove a unit from a loaded system
}
\examples{
\dontrun{
ud.remove.unit("miles")
}
}
\author{
Swechhya Bista \email{Swechhya.Bista@cytel.com}
}

