% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ud.functions.R
\name{ud.set.conversion}
\alias{ud.set.conversion}
\title{Set a conversion rule between two units}
\usage{
ud.set.conversion(unit1.text, unit2.text, system.name = "default",
  function.name = c("scale", "offset", "log", "invert"))
}
\arguments{
\item{unit1.text, unit2.text}{The text to define the units.  The text 
may take the form of "optional blank, optional floating point 
nonnegative number, optional blank, unit text, optional blank". The
blanks are ignored; if the floating point number is not given, then
it is assumed to be \code{1}.}

\item{system.name}{The unit system name to use.}

\item{function.name}{The function to use to convert between 
\code{unit1.text} and \code{unit2.text}.  See the details for more 
information.}
}
\value{
\code{TRUE} if conversion setting is successful, otherwise 
  returns an error.
}
\description{
Set a conversion rule between two units
}
\details{
The functions are applied so that if the units are given as
  \code{unit1.text = "a x"} and \code{unit2.text = "b y"} where "a"
  and "b" are numbers:
  \itemize{
    \item{scale}{x = b/a * y}
    \item{offset}{x = (b - a) * y}
    \item{log}{x = log(a/b)(y); log(a/b) is log base a/b}
    \item{invert}{x = 1/y}
  }

  If both unit text names are not defined, then both are created and
  they are mapped to each other.  If one is not defined, it is created
  and mapped to the other.  If both are defined, then an error is raised.
}
\examples{
\dontrun{
ud.set.conversion("1 moleaspirin", "180.157 gram") # Convert grams to moles of aspirin
}
}
\author{
Swechhya Bista \email{Swechhya.Bista@cytel.com}
}

